requirements.txt
Flask==2.3.3
Flask-OAuthlib==0.9.6
requests==2.31.0
requests-oauthlib==1.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7 